+++ package.json
@@ -49,7 +49,7 @@
     "cmdk": "^1.0.0",
     "date-fns": "^3.6.0",
     "embla-carousel-react": "^8.3.0",
-    "eslint": "8.49.0",
+    "eslint": "9.38.0",
     "eslint-config-next": "13.5.1",
     "framer-motion": "^10.x",
     "input-otp": "^1.2.4",
